Remove the old API directory
