--- ui/src/api/index.js
@@ -1,9 +1,0 @@
-import * as kapture from './kapture';
-import * as aws from './aws';
-
-export const api = {
-  kapture,
-  aws,
-};
-
-export { kapture, aws };
